refactored remaining js to React, tweaked clipboard copy style, added comments
