+++ _includes/header.html
@@ -16,7 +16,6 @@
 
     <!-- JS Includes -->
     <script src="https://code.jquery.com/jquery-2.2.4.min.js" integrity="sha256-BbhdlvQf/xTY9gja0Dq3HiwQF8LaCRTXxZKRutelT44=" crossorigin="anonymous"></script>
-    <script type="text/javascript" src="/js/main.js"></script>
 
     <!-- Google Analytics -->
     <script>
