+++ css/main.scss
@@ -113,6 +113,7 @@ header {
 .copyToClipboard{
     cursor: pointer;
     opacity: 0.7;
+	padding: 0 0.2em;
     &:hover {
         opacity: 1.0;
     }
