+++ src/EmailListItem.tsx
@@ -1,7 +1,5 @@
 ﻿import * as React from 'react';
 
-declare function copyToClipboard(element: any, link: any, t: any): any;
-
 export interface EmailListItemProps {
 	country: string,
 	state: string,
@@ -12,9 +10,36 @@ export interface EmailListItemProps {
 	regionId?: number
 };
 
-export class EmailListItem extends React.Component<EmailListItemProps> {
+interface EmailListItemState {
+	clickActive: boolean;
+}
+
+/**
+ * An individual email link, including copy button.
+ */
+export class EmailListItem extends React.Component<EmailListItemProps, EmailListItemState> {
+	copyButtonText: string = '🔗';
 	constructor(props: EmailListItemProps) {
 		super(props);
+		this.state = {
+			clickActive: false
+        }
+	}
+
+	copyToClipboard(copyText: string, isPermalink: boolean) {
+		this.setState({ clickActive: true });
+		setTimeout(() => this.setState({ clickActive: false }), 2000);
+
+		const element = document.createElement("textarea");
+		let copyValue = copyText;
+		if (isPermalink) {
+			copyValue = "https://defund12.org".concat(copyText);
+		}
+		element.value = copyValue;
+		document.body.appendChild(element);
+		element.select();
+		document.execCommand("copy");
+		document.body.removeChild(element);
 	}
 
 	render() {
@@ -22,7 +47,7 @@ export class EmailListItem extends React.Component<EmailListItemProps> {
 		return (
 			<li data-state={this.props.state}>
 				<a href={`${this.props.permalink}?browse`}>{this.props.city} - <i>{this.props.name}</i></a>
-				<span className="copyToClipboard" onClick={function (this: HTMLElement) { copyToClipboard(this, self.props.permalink, true) }}>🔗</span>
+				<span className="copyToClipboard" onClick={() => this.copyToClipboard(self.props.permalink, true)}>{(this.state.clickActive ? '✅(copied)' : '🔗')}</span>
 			</li>
 		);
 	}
